基于IonsCount的尝试代码告一段落,无法使用IonsCount的峰直接进行选峰鉴定,还需要更多的尝试

+++ src/main/java/net/csibio/propro/domain/bean/score/PeakGroup.java
@@ -17,7 +17,7 @@ public class PeakGroup {
     double apexRt;
 
     //最接近apexRt的光谱rt值
-    double nearestRt;
+    float nearestRt;
 
     //peak group中每一个碎片的强度
     HashMap<String, Double> ionIntensity;
